requirements.txt
Django>=4.2
djangorestframework>=3.14
python-dotenv>=1.0.1
pytz>=2024.1
drf-spectacular>=0.27.0
